requirements.txt
absl-py==1.0.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.12
easy-torch==1.2.10
easydict==1.9
google-auth==2.6.6
google-auth-oauthlib==0.4.6
grpcio==1.46.1
idna==3.3
importlib-metadata==4.11.3
setuptools==59.5.0
joblib==1.1.0
Markdown==3.3.7
numpy==1.22.3
oauthlib==3.2.0
Pillow==9.1.0
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.1.1
scipy==1.8.0
setproctitle==1.2.3
six==1.16.0
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
timm==0.6.7